requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
Pillow==10.4.0
aiohttp==3.10.10
aiofiles==24.1.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
openai==1.54.4
httpx==0.27.2
tenacity==9.0.0
loguru==0.7.2
